+++ Trains3D/Engine/Kernel/EGScene.m
@@ -35,7 +35,7 @@
     } withStart:@NO]);
 }
 
-- (void)updateWithDelta:(double)delta {
+- (void)updateWithDelta:(float)delta {
     [_controller updateWithDelta:delta];
 }
